+++ client/app/modules/login/login.component.js
@@ -4,10 +4,11 @@ import { ForgotPasswordController } from './forgotpassword.controller';
 
 export class LoginComponent {
   /*@ngInject*/
-  constructor(Auth, $location, $uibModal) {
-    this.Auth = Auth;
+  constructor($location, $window, $uibModal, Auth) {
     this.$location = $location;
+    this.$window = $window;
     this.$uibModal = $uibModal;
+    this.authService = Auth;
   }
 
   $onInit() {
@@ -16,9 +17,7 @@ export class LoginComponent {
       email: '',
       password: ''
     };
-    this.errors = {
-      login: undefined
-    };
+    this.errors = { login: undefined };
     this.submitted = false;
   }
 
@@ -26,10 +25,11 @@ export class LoginComponent {
     this.submitted = true;
 
     if(form.$valid) {
-      this.Auth.login({
-        email: this.user.email,
-        password: this.user.password
-      })
+      const { email, password } = this.user;
+      this.authService.login({ email, password })
+        .then(() => {
+          this.$window.history.back(); // usually login opens due to redirect
+        })
         .catch(err => {
           this.errors.login = err.message;
         });